+++ packages/theme/src/components/BlogRecommendArticle.vue
@@ -21,6 +21,10 @@ const recommendPadding = computed(() =>
 const recommend = computed(() =>
   _recommend === false ? undefined : _recommend
 )
+
+const showDate = computed(() => (_recommend && _recommend?.showDate) ?? true)
+const showNum = computed(() => (_recommend && _recommend?.showNum) ?? true)
+
 const title = computed(() => recommend.value?.title ?? (`<span class="svg-icon">${recommendSVG}</span>` + '相关文章'))
 const pageSize = computed(() => recommend.value?.pageSize || 9)
 const nextText = computed(() => recommend.value?.nextText || '换一组')
@@ -169,10 +173,16 @@ function handleLinkClick(link: string) {
       </ElButton>
     </div>
     <!-- 文章列表 -->
-    <ol v-if="currentWikiData.length" class="recommend-container">
-      <li v-for="(v, idx) in currentWikiData" :key="v.route">
+    <ol
+      v-if="currentWikiData.length" :class="{
+        'hide-num': !showNum,
+      }" class="recommend-container"
+    >
+      <li
+        v-for="(v, idx) in currentWikiData" :key="v.route"
+      >
         <!-- 序号 -->
-        <i class="num">{{ startIdx + idx + 1 }}</i>
+        <i v-if="showNum" class="num">{{ startIdx + idx + 1 }}</i>
         <!-- 简介 -->
         <div class="des">
           <!-- title -->
@@ -189,7 +199,7 @@ function handleLinkClick(link: string) {
             <span>{{ v.meta.title }}</span>
           </a>
           <!-- 描述信息 -->
-          <div class="suffix">
+          <div v-if="showDate" class="suffix">
             <!-- 日期 -->
             <span class="tag">{{ formatShowDate(v.meta.date) }}</span>
           </div>
@@ -234,6 +244,9 @@ function handleLinkClick(link: string) {
   padding: 0 10px 0 0px;
   width: 100%;
 
+  &.hide-num>li{
+    padding: 5px 0;
+  }
   li {
     display: flex;
 
@@ -253,6 +266,9 @@ function handleLinkClick(link: string) {
       overflow: hidden;
       text-overflow: ellipsis;
       white-space: nowrap;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
     }
 
     .title {
@@ -263,21 +279,21 @@ function handleLinkClick(link: string) {
       font-weight: 500;
       position: relative;
       cursor: pointer;
-
-      &.current {
+      transition: color .25s;
+      &.current,&:hover {
         color: var(--vp-c-brand-1);
       }
     }
 
-    .title:hover::after {
-      content: "";
-      position: absolute;
-      left: 0;
-      right: 0;
-      height: 0;
-      bottom: -3px;
-      border-bottom: 1px solid #b1b3b8;
-    }
+    // .title:hover::after {
+    //   content: "";
+    //   position: absolute;
+    //   left: 0;
+    //   right: 0;
+    //   height: 0;
+    //   bottom: -3px;
+    //   border-bottom: 1px solid #b1b3b8;
+    // }
 
     .suffix {
       font-size: 12px;